Add a simpler middleware, SynchronousMiddleware, that runs the phantomjs command synchronously and doesn't use any of the fancy polling that Shrimp::Middleware uses.  It also avoids showing the user an ugly "Preparing pdf..." page, which doesn't even go away after the PDF file has been downloaded.
The caveat is that you must make sure that your web server is able to process any number of
concurrent requests (or at least enough concurrent requests that you aren't worried).  If your web
server reaches its concurrency limit and then receives a request for a .pdf resource, then when
phantomjs attempts to request the .html version of the resource, this new HTTP request will have to
wait until the server finishes some other request and is able to serve this new HTTP request.

That should generally be okay as long as most of the other requests are quick enough.  But in the
worst case, where each of the requests currently being served is a request for a .pdf resource
(which must generate an additional request for the .html request), then it will create a deadlock,
because request A blocks while waiting for request B to complete, while at the same time request B
can't complete because it is blocking while waiting for request A to finish (in the simple case
where the max concurrency is 2).

In Rails, if you are using a multi-threaded web server such as puma, then you *must* set
config.allow_concurrency = true to allow concurrent requests.  Otherwise, the Rack::Lock middleware
will prevent the 2nd request (that PhantomJS generates) from being processed until the 1st request
(for the .pdf file) has completed, which immediately creates a deadlock!